requirements.txt
flask
flask-cors
biopython
matplotlib
gunicorn
ghp-import
